refactor(sources): introduce BasePaperSource registry, one module per source
Stateless `BasePaperSource` subclass per paper source under
`src/shared/js/utils/sources/` plus a central registry
(`sources/index.js`) exposing `getSource`, `allSources`, `matchUrl`,
`sourceFromIs`, `preprintSources`, and a `knownPaperPages` shim.

`paper.js`, `urls.js`, `parsers.js`, `data.js`, and `functions.js` now
dispatch through the registry instead of large per-source switch / if
chains. `parsers.js` keeps only shared HTTP/DOM/BibTeX helpers; the
`make<Source>Paper` family is gone. `preprintMatching.js` extracts
`tryPWCMatch`, `tryPreprintMatch`, `tryCrossRef`, `tryDBLP`,
`trySemanticScholar`, `tryGoogleScholar`, `tryUnpaywall`, and
`decodeHtml` from the old `parsers.js`. `initPaper` and `autoTagPaper`
move to `paper.js` as lifecycle (not parsing) helpers.

Bundled cleanups:

- `getDisplayId(paper)` (was `getDisplayId(id)`) — removes a
  `state.papers` global read; nature/acs/iop suffixes flow through
  `BasePaperSource.displayId`.
- `sourceExtras.springer.types`, `overrideORConfs`,
  `overridePMLRConfs` colocated on their respective source classes.
- `preprintSources` array replaced by `static isPreprint = true` on
  arxiv/biorxiv; registry exposes `preprintSources()`.
- Drop unreachable `default: xkcd/1969` branches in
  `paperToAbs`/`paperToPDF`.
- `config.js` keeps DBLP overrides only and re-exports
  `knownPaperPages` / `preprintSources` from the registry for
  back-compat consumers (e.g. `PMUtils.config.knownPaperPages`).
- `PMDebug` exposes `sources` and `preprintMatching` modules.

Made-with: Cursor
