+++ src/shared/js/utils/config.js
@@ -149,246 +149,6 @@ export const prefsCheckDefaultFalse = [
  */
 export const prefsStorageKeys = [...prefsCheckNames, "pdfTitleFn"];
 
-/**
- * Extra data per source
- */
-export const sourceExtras = {
-    springer: {
-        types: ["chapter", "article", "book", "referenceworkentry"],
-    },
-};
-
-/**
- * Sources which are preprints (important for de-duplication)
- */
-export const preprintSources = ["arxiv", "biorxiv"];
-
-/**
- * Map of known data sources to the associated paper urls: pdf urls and web-pages urls.
- * IMPORTANT: paper page before pdf (see background script)
- * Notes:
- *  ijcai -> papers < 2015 will not be parsed due to website changes
- *           (open an issue if that's problematic)
- */
-export const knownPaperPages = {
-    acl: {
-        patterns: ["aclanthology.org/"],
-        name: "ACL Anthology (Association for Computational Linguistics)",
-    },
-    acm: {
-        patterns: ["dl.acm.org/doi/"],
-        name: "ACM (Association for Computing Machinery)",
-    },
-    aps: {
-        patterns: [
-            (url) => Boolean(url.match(/journals\.aps\.org\/\w+\/(abstract|pdf)\//g)),
-        ],
-        name: "APS (American Physical Society)",
-    },
-    acs: {
-        patterns: ["pubs.acs.org/doi/"],
-        name: "ACS (American Chemical Society)",
-    },
-    arxiv: {
-        patterns: [
-            "arxiv.org/abs/",
-            "arxiv.org/pdf/",
-            "scirate.com/arxiv/",
-            "ar5iv.labs.arxiv.org/html/",
-            "alphaxiv.org/abs/",
-            "alphaxiv.org/pdf/",
-            (url) =>
-                url.includes("huggingface.co/papers/") &&
-                url.split("huggingface.co/papers/")[1].match(/\d+\.\d+/),
-        ],
-        name: "ArXiv",
-    },
-    biorxiv: {
-        patterns: ["biorxiv.org/content"],
-        name: "BioRxiv",
-    },
-    cell: {
-        patterns: [
-            (url) =>
-                url.includes("cell.com/") &&
-                url.split("cell.com/")[1].match(/\d{4}-\d{3}[0-9X]/),
-        ],
-        name: "Cell",
-    },
-    chemrxiv: {
-        patterns: [
-            "chemrxiv.org/engage/chemrxiv/article-details/",
-            "https://chemrxiv.org/doi/",
-            (url) =>
-                url.includes(
-                    "https://chemrxiv.org/engage/api-gateway/chemrxiv/assets",
-                ) && url.endsWith(".pdf"),
-        ],
-        name: "ChemRxiv",
-    },
-    cvf: {
-        patterns: ["openaccess.thecvf.com/content"],
-        name: "CVF (Computer Vision Foundation)",
-    },
-    frontiers: {
-        patterns: [
-            "frontiersin.org/articles",
-            (url) => url.match(/frontiersin\.org\/.+\/articles\//),
-        ],
-        name: "Frontiers",
-    },
-    hal: {
-        patterns: [
-            (url) => /hal\.science\/\w+-\d+(v\d+)?(\/document)?$/gi.test(url),
-            (url) => /hal\.science\/\w+-\d+(v\d+)?\/file\/.+\.pdf$/gi.test(url),
-        ],
-        name: "HAL",
-    },
-    ihep: {
-        patterns: ["inspirehep.net/literature/", "inspirehep.net/files/"],
-        name: "IHEP (INSPIRE - High Energy Physics)",
-    },
-    ijcai: {
-        patterns: [(url) => /ijcai\.org\/proceedings\/\d{4}\/\d+/gi.test(url)],
-        name: "IJCAI (International Joint Conferences on Artificial Intelligence)",
-    },
-    ieee: {
-        patterns: [
-            "ieeexplore.ieee.org/document/",
-            "ieeexplore.ieee.org/abstract/document/",
-            "ieeexplore.ieee.org/stamp/stamp.jsp?tp=&arnumber=",
-        ],
-        name: "IEEE (Institute of Electrical and Electronics Engineers)",
-    },
-    iop: {
-        patterns: ["iopscience.iop.org/article/"],
-        name: "IOP (Institute Of Physics)",
-    },
-    jmlr: {
-        patterns: [(url) => url.includes("jmlr.org/papers/v") && !url.endsWith("/")],
-        name: "JMLR (Journal of Machine Learning Research)",
-    },
-    mdpi: {
-        patterns: [(url) => /mdpi\.com\/\d+-.+/gi.test(url)],
-        name: "MDPI (Multidisciplinary Digital Publishing Institute)",
-    },
-    nature: {
-        patterns: ["nature.com/articles/"],
-        name: "Nature",
-    },
-    neurips: {
-        patterns: [
-            "neurips.cc/paper/",
-            "neurips.cc/paper_files/paper/",
-            "nips.cc/paper/",
-            "nips.cc/paper_files/paper/",
-        ],
-        name: "NeurIPS (Neural Information Processing Systems)",
-    },
-    openreview: {
-        patterns: [
-            "openreview.net/forum",
-            "openreview.net/pdf",
-            "openreview.net/attachment",
-        ],
-        name: "OpenReview",
-    },
-    oup: {
-        patterns: [
-            (url) =>
-                (url
-                    .split("https://academic.oup.com/")[1]
-                    ?.split("/")[1]
-                    ?.indexOf("article") ?? -1) >= 0,
-        ],
-        name: "OUP (Oxford University Press)",
-    },
-    plos: {
-        patterns: [(url) => /journals\.plos\.org\/.+\/article.+id=/gi.test(url)],
-        name: "PLOS (Public Library of Science)",
-    },
-    pmc: {
-        patterns: [
-            "ncbi.nlm.nih.gov/pmc/articles/PMC",
-            "ncbi.nlm.nih.gov/articles/PMC",
-            (url) => url.match(/ncbi.nlm.nih.gov\/\d+/),
-        ],
-        name: "PMC (PubMed Central)",
-    },
-    pmlr: {
-        patterns: ["proceedings.mlr.press/"],
-        name: "PMLR (Proceedings of Machine Learning Research)",
-    },
-    pnas: {
-        patterns: ["pnas.org/content/", "pnas.org/doi/"],
-        name: "PNAS (Proceedings of the National Academy of Sciences)",
-    },
-    rsc: {
-        patterns: ["pubs.rsc.org/en/content/article"],
-        name: "RSC (Royal Society of Chemistry)",
-    },
-    science: {
-        patterns: [
-            (url) => Boolean(url.match(/science\.org\/doi\/?(abs|full|pdf|epdf)?\//g)),
-        ],
-        name: "Science",
-    },
-    sciencedirect: {
-        patterns: [
-            "sciencedirect.com/science/article/pii/",
-            "sciencedirect.com/science/article/abs/pii/",
-            "reader.elsevier.com/reader/sd/pii/",
-        ],
-        name: "ScienceDirect",
-    },
-    springer: {
-        patterns: [
-            ...sourceExtras.springer.types.map((type) => `link.springer.com/${type}/`),
-            "link.springer.com/content/pdf/",
-        ],
-        name: "Springer",
-    },
-    website: {
-        // special case, manual parsing of arbitrary websites
-        patterns: [],
-        name: "Manually parsed website",
-    },
-    wiley: {
-        patterns: [
-            (url) =>
-                Boolean(
-                    url.match(
-                        /onlinelibrary\.wiley\.com\/doi\/(abs\/|full\/|pdf\/|epdf\/|10\.)/g,
-                    ),
-                ),
-        ],
-        name: "Wiley",
-    },
-    aip: {
-        patterns: [
-            (url) =>
-                Boolean(
-                    url.match(
-                        /pubs.aip.org\/aip\/.+\/(article|article-abstract|article-split)\//g,
-                    ) || url.match(/watermark.silverchair.com\/.+\.pdf/g),
-                ),
-        ],
-        name: "AIP (American Institute of Physics)",
-    },
-};
-
-export const overrideORConfs = {
-    "robot-learning": "CoRL",
-    ijcai: "IJCAI",
-};
-export const overridePMLRConfs = {
-    "Conference on Learning Theory": "CoLT",
-    "International Conference on Machine Learning": "ICML",
-    "Conference on Uncertainty in Artificial Intelligence": "UAI",
-    "Conference on Robot Learning": "CoRL",
-    "International Conference on Artificial Intelligence and Statistics": "AISTATS",
-    "International Conference on Algorithmic Learning Theory": "ALT",
-};
 export const overrideDBLPVenues = {
     "J. Mach. Learn. Res.": "JMLR",
 };
@@ -537,3 +297,5 @@ export const englishStopWords = new Set([
 ]);
 
 export const journalAbbreviations = {};
+
+export { knownPaperPages, preprintSources } from "./sources/index.js";
