+++ src/shared/js/utils/parsers.js
@@ -3,63 +3,28 @@ import {
     log,
     info,
     warn,
-    cleanStr,
     miniHash,
     sendMessageToBackground,
-    cleanBiorxivURL,
-    parseCVFUrl,
     firstNonStopLowercase,
-    urlToWebsiteId,
     isPdfUrl,
     noParamUrl,
-    dedent,
     parseUrl,
-    arxivIdFromURL,
     spaceCamelCase,
     logError,
-    silentPromiseTimeout,
     toSingleSpace,
+    parseCVFUrl,
 } from "@pmu/functions.js";
-import {
-    state,
-    preprintSources,
-    overrideORConfs,
-    overridePMLRConfs,
-    overrideDBLPVenues,
-    englishStopWords,
-    sourceExtras,
-    journalAbbreviations,
-} from "@pmu/config.js";
-import { getStorage, setStorage, validatePaper, makeVenue } from "@pmu/data.js";
+import { state, journalAbbreviations } from "@pmu/config.js";
 import {
     extractBibtexValue,
     bibtexToString,
     bibtexToObject,
-    sanitizeBibtexObject,
 } from "@pmu/bibtexParser.js";
 import { queryAll, querySelector } from "@pmu/miniquery.js";
-import { sleep } from "@pmu/sync.js";
-import { parseIdFromUrl } from "@pmu/urls.js";
 import { readJournalAbbreviations } from "@pmu/state.js";
-// -------------------
-// -----  Utils  -----
-// -------------------
-
-export const decodeHtml = (html) => {
-    // https://stackoverflow.com/questions/5796718/html-entity-decode
-    var txt = document.createElement("textarea");
-    txt.innerHTML = html;
-    return txt.value;
-};
-
 export const flipAuthor = (author) => author.split(", ").reverse().join(" ");
 export const flipAndAuthors = (authors) =>
     authors.split(" and ").map(flipAuthor).join(" and ");
-
-// -------------------
-// -----  Fetch  -----
-// -------------------
-
 export const fetchCvfHTML = async (url) => {
     let paperPage, text;
     if (url.endsWith(".pdf")) {
@@ -404,148 +369,6 @@ export const extractDataFromDCMetaTags = (dom) => {
     };
 };
 
-export const makeArxivPaper = async (url) => {
-    const arxivId = arxivIdFromURL(url);
-    const xmlData = await sendMessageToBackground({
-        type: "fetch-arxiv-xml",
-        paperId: arxivId,
-    });
-    const doc = new DOMParser().parseFromString(
-        xmlData.replaceAll("\n", ""),
-        "text/xml",
-    );
-
-    const authors = queryAll("author name", doc).map((el) => el.innerHTML);
-    const author = authors.join(" and ");
-
-    const pdfLink = [...doc.getElementsByTagName("link")]
-        .map((l) => l.getAttribute("href"))
-        .filter((h) => h.includes("arxiv.org/pdf/"))[0]
-        .replace(/v\d+(\.pdf)?$/gi, ".pdf");
-
-    let title = doc.querySelector("entry title");
-    title = title?.textContent || title?.innerText || "";
-    const year = doc.querySelector("entry published").innerHTML.slice(0, 4);
-    const key =
-        authors[0].split(" ").last().toLowerCase() +
-        year +
-        firstNonStopLowercase(title);
-
-    const id = `Arxiv-${arxivId.replace("/", "_")}`;
-
-    let bibtex = "";
-    bibtex += `@article{${key},\n`;
-    bibtex += `    title={${title} },\n`;
-    bibtex += `    author={${author} },\n`;
-    bibtex += `    year={${year}},\n`;
-    bibtex += `    journal={arXiv preprint arXiv: ${arxivId}}\n`;
-    bibtex += `}`;
-
-    const venue = "";
-
-    return { author, bibtex, id, key, pdfLink, title, venue, year };
-};
-
-export const makeNeuripsPaper = async (url) => {
-    if (url.endsWith(".pdf")) {
-        url = url
-            .replace("/file/", "/hash/")
-            .replace(/-Paper(.*)\.pdf/, "-Abstract$1.html");
-    }
-    const hash = url.split("/").slice(-1)[0].split("-Paper")[0];
-
-    const dom = await fetchDom(url);
-
-    const citeUrl = [...dom.getElementsByTagName("a")]
-        .filter((a) => a.innerText === "Bibtex")[0]
-        ?.getAttribute("href");
-
-    let bibtex, author, title, year, key, citationKey;
-
-    if (citeUrl) {
-        bibtex = await fetchText(`https://${parseUrl(url).host}${citeUrl}`);
-        ({ author, citationKey, title, year } = bibtexToObject(bibtex));
-        author = flipAndAuthors(author);
-        key = citationKey;
-    } else {
-        const paragraphs = queryAll(".container-fluid .col p", dom);
-
-        title = dom.getElementsByTagName("h4")[0].innerHTML;
-        const h4Authors = queryAll("h4", dom).filter(
-            (h) => h.innerText === "Authors",
-        )[0];
-
-        author = h4Authors.nextElementSibling.innerText
-            .split(", ")
-            .map((author) =>
-                author
-                    .split(" ")
-                    .map((p) => p.capitalize())
-                    .join(" "),
-            )
-            .join(" and ");
-        year = paragraphs[0].innerHTML.match(/\d{4}/)[0];
-        key = `neurips${year}${hash.slice(0, 8)}`;
-
-        bibtex = "";
-        bibtex += `@inproceedings{NEURIPS${year}_${hash.slice(0, 8)},\n`;
-        bibtex += `    author={${author}},\n`;
-        bibtex += `    booktitle={Advances in Neural Information Processing Systems},\n`;
-        bibtex += `    editor={H.Larochelle and M.Ranzato and R.Hadsell and M.F.Balcan and H.Lin},\n`;
-        bibtex += `    publisher={Curran Associates, Inc.},\n`;
-        bibtex += `    title={${title}},\n`;
-        bibtex += `    url={${url}},\n`;
-        bibtex += `    year={${year}}\n`;
-        bibtex += `}`;
-        bibtex = bibtexToString(bibtex);
-    }
-
-    const pdfLink = url
-        .replace("/hash/", "/file/")
-        .replace("-Abstract.html", "-Paper.pdf");
-    const id = `NeurIPS-${year}_${hash.slice(0, 8)}`;
-    const venue = "NeurIPS";
-    const note = `Accepted @ ${venue} (${year})`;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeCVFPaper = async (url) => {
-    const htmlText = await fetchCvfHTML(url);
-    const dom = new DOMParser().parseFromString(
-        htmlText.replaceAll("\n", ""),
-        "text/html",
-    );
-    const title = dom.getElementById("papertitle").innerText.trim();
-    let author = dom
-        .querySelector("#authors i")
-        .innerText.split(",")
-        .map((a) => a.trim())
-        .join(" and ");
-    const { year, id, conf } = parseCVFUrl(url);
-    let pdfLink = "";
-    if (url.endsWith(".pdf")) {
-        pdfLink = url;
-    } else {
-        let href = [...dom.getElementsByTagName("a")]
-            .filter((a) => a.innerText === "pdf")[0]
-            .getAttribute("href");
-        if (href.startsWith("../")) {
-            href = href.replaceAll("../", "");
-        }
-        if (!href.startsWith("/")) {
-            href = "/" + href;
-        }
-        pdfLink = "http://openaccess.thecvf.com" + href;
-    }
-    const venue = conf;
-    const note = `Accepted @ ${venue} (${year})`;
-    const bibtex = bibtexToString(dom.querySelector(".bibref").innerText);
-    const key = bibtex.split("{")[1].split(",")[0];
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
 export const makeOpenReviewBibTex = (paper, url) => {
     const title = paper.content.title;
     const author = paper.content.authors.join(" and ");
@@ -591,833 +414,10 @@ export const extractAPIv2ContentValue = (paper) => {
     paper.content = content;
     return { isV2, paper };
 };
-
-export const makeOpenReviewPaper = async (url) => {
-    const noteJson = await getOpenReviewNoteJSON(url);
-    const forumJson = await getOpenReviewForumJSON(url);
-
-    if (noteJson.status === 403 && noteJson.name === "ForbiddenError") {
-        logError(
-            dedent(`Error parsing OpenReview url ${url}.
-            Most likely because this entry is protected and you do not have the rights to access it.
-
-            1/ Make sure you are logged in.
-            2/ Alternatively, this may be due to OpenReview changing the visibility of this paper.
-
-            Try accessing this URL manually to make sure.`),
-        );
-        throw Error(noteJson.message);
-    } else if (noteJson.status === 404 && noteJson.name === "NotFoundError") {
-        logError(dedent(`Error parsing OpenReview url ${url}.`));
-        throw Error(noteJson.message);
-    }
-
-    var paper = noteJson.notes[0];
-    let isV2 = false;
-    var forum = forumJson.notes;
-
-    ({ isV2, paper } = extractAPIv2ContentValue(paper));
-    console.log("paper", paper);
-    const title = paper.content.title;
-    const author = (
-        paper.content.authors ||
-        paper.content.authors?.value || ["Anonymous"]
-    ).join(" and ");
-    const bibtex = bibtexToString(
-        paper.content._bibtex || makeOpenReviewBibTex(paper, url),
-    );
-    const bibObj = bibtexToObject(bibtex);
-    const key = bibObj.citationKey;
-    const year = bibObj.year;
-
-    let pdfLink;
-    if (paper.pdf) {
-        pdfLink = `https://openreview.net/pdf?id=${paper.id}`;
-    } else {
-        if (paper.html) {
-            pdfLink = paper.html.replace("/forum?id=", "/pdf?id=");
-        } else {
-            pdfLink = url.replace("/forum?id=", "/pdf?id=");
-        }
-    }
-
-    const confParts = paper.invitation?.split("/") || paper.domain.split("/");
-    let organizer = confParts[0].split(".")[0];
-    let event = confParts
-        .slice(1)
-        .join("/")
-        .split("-")[0]
-        .replaceAll("/", " ")
-        .replace(" Conference", "");
-
-    let overrideOrg = organizer;
-    let overridden = false;
-    if (overrideORConfs.hasOwnProperty(organizer)) {
-        overrideOrg = overrideORConfs[organizer];
-        overridden = true;
-    }
-    if (overridden) {
-        event = event.replace(overrideOrg, "");
-        organizer = overrideOrg;
-    }
-
-    const conf = `${organizer} ${event}`
-        .replace(/ \d\d\d\d/g, "")
-        .replace(/\s\s+/g, " ");
-    const id = `OR-${organizer}-${year}_${paper.id}`;
-
-    let candidates, decision, note;
-
-    candidates = isV2
-        ? forum.filter((r) => r?.content?.recommendation?.value)
-        : forum.filter(
-              (r) =>
-                  ["Final Decision", "Paper Decision", "Acceptance Decision"].indexOf(
-                      r?.content?.title,
-                  ) > -1,
-          );
-    let venue = "";
-    if (candidates && candidates.length > 0) {
-        decision = isV2
-            ? candidates[0].content.recommendation.value
-            : candidates[0].content.decision;
-        decision = decision
-            .split(" ")
-            .map((v, i) => {
-                return i === 0 ? cleanStr(v) + "ed" : v;
-            })
-            .join(" ");
-        note = `${decision} @ ${conf} (${year})`;
-        if (decision.toLowerCase().indexOf("rejected") < 0) {
-            venue = conf;
-        }
-    }
-
-    if (author !== "Anonymous" && !venue && bibObj.booktitle) {
-        note = `Accepted @ ${bibObj.booktitle}`;
-        venue = bibObj.booktitle;
-    }
-    if (author === "Anonymous" && decision != "Rejected") {
-        note = `Under review @ ${conf} (${year}) (${new Date().toLocaleDateString()})`;
-    }
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeBioRxivPaper = async (url) => {
-    let author, bibtex, id, key, note, pdfLink, title, venue, year;
-    const biorxivAPI = "https://api.biorxiv.org";
-    const pageURL = url.replace(".full.pdf", "");
-    let doi = url.split("/").slice(-2).join("/").replace(".full.pdf", "").split("v")[0];
-    const api = `${biorxivAPI}/details/biorxiv/${doi}`;
-    const data = await fetch(api).then((response) => {
-        return response.json();
-    });
-
-    if (data.messages[0].status !== "ok")
-        throw new Error(`${api} returned ${data.messages[0].status}`);
-    const paper = data.collection.last();
-
-    if (paper.published.startsWith("10.")) {
-        doi = paper.published;
-        const paperData = await fetchBibtexToPaper({ doi });
-        ({ author, bibtex, key, note, title, venue, year } = paperData);
-    } else {
-        const pageText = await fetchText(pageURL);
-
-        const dom = new DOMParser().parseFromString(
-            pageText.replaceAll("\n", ""),
-            "text/html",
-        );
-        const bibtextLink = dom.querySelector(".bibtext a").getAttribute("href");
-
-        bibtex = bibtexToString(await (await fetch(bibtextLink)).text());
-        author = extractAuthor(bibtex);
-
-        key = bibtex.split("\n")[0].split("{")[1].replace(",", "").trim();
-        note = "";
-        title = paper.title;
-        year = paper.date.split("-")[0];
-        venue = "";
-    }
-    pdfLink = cleanBiorxivURL(url) + ".full.pdf";
-    id = await parseIdFromUrl(url);
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makePMLRPaper = async (url) => {
-    const key = url.split("/").last().split(".")[0];
-    const id = await parseIdFromUrl(url);
-
-    const absURL = url.includes(".html")
-        ? url
-        : url.split("/").slice(0, -2).join("/") + `/${key}.html`;
-
-    const pdfLink = absURL.replace(".html", "") + `/${key}.pdf`;
-
-    const dom = await fetchDom(absURL);
-
-    const bibtexRaw = dom
-        .getElementById("bibtex")
-        .innerText.replaceAll("\t", " ")
-        .replaceAll(/\s\s+/g, " ");
-    let bibtex = bibtexRaw;
-    const items = bibtexRaw.match(/,\ ?\w+ ?= ?{/g);
-    for (const item of items) {
-        bibtex = bibtex.replace(
-            item,
-            item.replace(", ", ",\n    ").replace(" = ", "="),
-        );
-    }
-    if (bibtex.endsWith("}}")) {
-        bibtex = bibtex.slice(0, -2) + "}\n}";
-    }
-    bibtex = bibtexToString(bibtex);
-
-    const author = extractAuthor(bibtex);
-    const title = dom.getElementsByTagName("h1")[0].innerText;
-    const year = extractBibtexValue(bibtex, "year");
-
-    let conf = extractBibtexValue(bibtex, "booktitle").replaceAll(
-        "Proceedings of the",
-        "",
-    );
-    let venue = conf;
-    let note = `Accepted @ ${venue} (${year})`;
-    for (const long in overridePMLRConfs) {
-        if (conf.includes(long)) {
-            venue = overridePMLRConfs[long];
-            conf = venue + " " + year;
-            note = "Accepted @ " + conf;
-            break;
-        }
-    }
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
 export const findACLValue = (dom, key) => {
     const dt = queryAll("dt", dom).filter((v) => v.innerText.includes(key))[0];
     return dt.nextElementSibling.innerText;
 };
-
-export const makeACLPaper = async (url) => {
-    url = url.replace(".pdf", "");
-    const dom = await fetchDom(url);
-
-    const bibtexEl = dom.getElementById("citeBibtexContent");
-    if (!bibtexEl) return;
-
-    const title = dom.getElementById("title").innerText;
-    const bibtex = bibtexToString(bibtexEl.innerText);
-
-    const bibtexData = bibtexToObject(bibtex);
-
-    const year = bibtexData.year;
-    const author = bibtexData.author
-        .replace(/\s+/g, " ")
-        .split(" and ")
-        .map((v) =>
-            v
-                .split(",")
-                .map((a) => a.trim())
-                .reverse()
-                .join(" "),
-        )
-        .join(" and ");
-    const key = bibtexData.citationKey;
-
-    const venue = findACLValue(dom, "Venue");
-    const pdfLink = findACLValue(dom, "PDF");
-    const aid = findACLValue(dom, "Anthology ID");
-
-    const id = `ACL-${venue}-${year}_${aid}`;
-    const note = `Accepted @ ${venue} (${year})`;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makePNASPaper = async (url) => {
-    /*
-        https://www.pnas.org/doi/10.1073/pnas.2114679118
-        https://www.pnas.org/doi/epdf/10.1073/pnas.2114679118
-        https://www.pnas.org/doi/pdf/10.1073/pnas.2114679118
-    */
-
-    url = url.replace(".full.pdf", "").replace(/\/doi\/e?pdf\//, "/doi/abs/");
-    const dom = await fetchDom(url);
-
-    const title = dom.getElementsByTagName("h1")[0].innerText;
-    const author = queryAll(
-        ".authors span[property='author'] a:not([property='email']):not(.orcid-id)",
-        dom,
-    )
-        .filter((el) => !el.getAttribute("href").includes("mailto:"))
-        .map((el) => el.innerText)
-        .join(" and ");
-
-    const year = dom
-        .querySelector("span[property='datePublished']")
-        .innerText.match(/\d{4}/g)[0];
-
-    const pid = url.endsWith("/")
-        ? url.split("/").slice(-2)[0]
-        : url.split("/").slice(-1)[0];
-
-    const id = `PNAS-${year}_${pid}`;
-    const pdfLink =
-        url.includes("/doi/pdf/") || url.includes("/doi/epdf/")
-            ? url.replace("/doi/epdf/", "/doi/pdf/")
-            : url.replace("/doi/abs/", "/doi/pdf/").replace("/doi/full/", "/doi/pdf/");
-    const doi = [...dom.querySelector(".core-container").getElementsByTagName("a")]
-        .map((a) => a.getAttribute("href"))
-        .filter((a) => a?.includes("https://doi.org"))[0]
-        .split("/")
-        .slice(-2)
-        .join("/");
-    const key = `doi:${doi}`;
-    const bibtex = bibtexToString(`
-    @article{${key},
-        author={${author}},
-        title={${title}},
-        journal = {Proceedings of the National Academy of Sciences},
-        year={${year}},
-        doi={${doi}},
-        eprint={${pdfLink}},
-        URL={${pdfLink.replace("/doi/pdf/", "/doi/abs/")}}
-    }`);
-    const venue = "PNAS";
-
-    const note = `Published @ PNAS (${year})`;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeNaturePaper = async (url) => {
-    url = url.replace(".pdf", "").split("#")[0];
-    const pdfLink = url + ".pdf";
-    const hash = url.split("/").last();
-
-    const dom = await fetchDom(url);
-
-    const title = dom.querySelector("h1.c-article-title").innerText;
-    const author = queryAll("ul.c-article-author-list li", dom)
-        .map((a) =>
-            a.innerText
-                .replace(/(\ ?,)|&|…|\d/g, "")
-                .split(/orcid/i)[0]
-                .trim(),
-        )
-        .filter((a) => a.length > 0)
-        .join(" and ");
-    const year = dom
-        .querySelector(".c-article-info-details")
-        .innerText.match(/\(\d{4}\)/)[0]
-        .replace(/\(|\)/g, "");
-    const journal = dom.querySelector(".c-article-info-details [data-test]").innerText;
-    const id = `Nature-${year}_${hash}`;
-
-    const doiClasses = [
-        ".c-bibliographic-information__citation",
-        ".c-bibliographic-information__value",
-    ];
-    let doi;
-    for (const doiClass of doiClasses) {
-        doi = querySelector(doiClass)?.innerText.split("https://doi.org/")[1];
-        if (doi) break;
-    }
-    if (!doi) {
-        doi = [...dom.getElementsByTagName("span")]
-            .map((a) => a.innerText)
-            .filter((a) => a.includes("https://doi.org"))[0];
-    }
-
-    const key = `${author.split(" ")[1]}${year}${firstNonStopLowercase(title)}`;
-    let bibData = {
-        citationKey: key,
-        entryType: "article",
-        author,
-        title,
-        journal,
-        year,
-    };
-    if (doi) {
-        bibData.doi = doi;
-        bibData.url = `https://doi.org/${doi}`;
-    }
-    const bibtex = bibtexToString(bibData);
-    const note = `Published @ ${journal} (${year})`;
-    const venue = journal;
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeACSPaper = async (url) => {
-    url = url.replace("pubs.acs.org/doi/pdf/", "pubs.acs.org/doi/").split("?")[0];
-    const doi = url.replace("/abs/", "/").split("/doi/")[1];
-    const citeUrl = `https://pubs.acs.org/action/downloadCitation?doi=${doi}&include=cit&format=bibtex&direct=true`;
-    const bibtex = await fetchText(citeUrl);
-    const data = bibtexToObject(bibtex);
-    const author = data.author.replaceAll("\n", "").trim();
-    const title = data.title.trim();
-    const year = data.year.trim();
-    const key = data.citationKey.trim();
-    const pdfLink = `https://pubs.acs.org/doi/pdf/${doi}`;
-    const note = `Published @ ${data.journal} (${data.year})`;
-    const id = `ACS_${doi.replaceAll(".", "").replaceAll("/", "")}`;
-    const venue = data.journal;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeIOPPaper = async (url) => {
-    let author, bibtex, id, key, note, pdfLink, title, venue, year;
-    url = url.split("#")[0];
-    if (url.endsWith("/pdf")) url = url.slice(0, -4);
-
-    const doi = url.split("/article/").last().split("/meta")[0];
-
-    const data = await fetchBibtexToPaper({ doi });
-
-    ({ author, bibtex, key, note, title, venue, year } = data);
-    id = `IOPscience_${miniHash(doi)}`;
-    pdfLink = url + "/pdf";
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeJMLRPaper = async (url) => {
-    if (url.includes("/papers/volume")) {
-        url = url.replace("/papers/volume", "/papers/v");
-    }
-    if (url.endsWith(".pdf")) {
-        url = url.split("/").slice(0, -1).join("/");
-    }
-    url = url.replace(".html", "");
-    const jid = url.split("/").last();
-    const citeUrl = url + ".bib";
-    const bibtex = await fetchText(citeUrl);
-    const data = bibtexToObject(bibtex);
-
-    const { author, year, title, citationKey } = data;
-    const key = citationKey.trim();
-    const id = `JMLR-${year}_${jid}`;
-    const note = `Published @ JMLR (${year})`;
-    const pdfLink = url.replace("/papers/v", "/papers/volume") + `/${jid}.pdf`;
-    const venue = "JMLR";
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makePMCPaper = async (url) => {
-    url = noParamUrl(url);
-    if (url.endsWith("/")) {
-        url = url.slice(0, -1);
-    }
-    if (isPdfUrl(url)) {
-        url = url.split("/pdf")[0];
-    }
-    const pmcid = url.includes("PMC")
-        ? url.match(/PMC\d+/)[0].replace("PMC", "")
-        : url.match(/ncbi.nlm.nih.gov\/(\d+)/)[1];
-    const pdfLink = url + "/pdf";
-    const html = await fetchText(url);
-    const doi = html.match(/10\.\d{4,9}\/[-._;()/:A-Z0-9]+/i)[0];
-    const { author, bibtex, key, note, title, venue, year } = await fetchBibtexToPaper({
-        doi,
-    });
-    const id = `PMC-${year}_${pmcid}`;
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeIJCAIPaper = async (url) => {
-    const procId = url.endsWith(".pdf")
-        ? url
-              .replace(".pdf", "")
-              .split("/")
-              .last()
-              .match(/[1-9]\d*/)[0]
-        : url.split("/").last();
-
-    const year = url.match(/proceedings\/\d+/gi)[0].split("/")[1];
-
-    // ijcai bibtexs have issues with note = {NOTE}\n with a missing ","
-    const bibtex = (
-        await fetchText(`https://www.ijcai.org/proceedings/${year}/bibtex/${procId}`)
-    ).replace(/}\n/gi, "},\n");
-    const data = bibtexToObject(
-        bibtex
-            .split("\n")
-            .filter((line) => !/note\s+=/gi.test(line))
-            .join("\n"),
-    );
-
-    const key = data.citationKey;
-    const title = data.title;
-    const author = flipAndAuthors(data.author);
-    const id = `IJCAI-${year}_${procId}`;
-    const note = `Accepted @ IJCAI (${year})`;
-    const venue = "IJCAI";
-    const pdfId = procId.padStart(4, 0);
-    const pdfLink = `https://www.ijcai.org/proceedings/${year}/${pdfId}.pdf`;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeACMPaper = async (url) => {
-    let author, year, title, venue, key, bibtex, note, id, doi, pdfLink;
-    url = noParamUrl(url);
-    if (isPdfUrl(url)) {
-        pdfLink = url;
-    } else {
-        pdfLink = url.replace(/\/doi\/?(abs|full)?\//, "/doi/pdf/");
-    }
-    doi = "10.5555/" + url.split("10.5555/")[1];
-    const response = await fetch("https://dl.acm.org/action/exportCiteProcCitation", {
-        headers: {
-            "content-type": "application/x-www-form-urlencoded; charset=UTF-8",
-        },
-        referrer: `https://dl.acm.org/doi/${doi}`,
-        body: `dois=${doi}&targetFile=custom-bibtex&format=bibTex`,
-        method: "POST",
-        mode: "cors",
-    });
-    if (response.ok) {
-        const data = await response.json();
-        if (data && data.items && data.items.length > 0) {
-            const item = data.items[0][doi];
-            title = item.title;
-            author = item.author.map((a) => `${a.given} ${a.family}`).join(" and ");
-            year = item.issued["date-parts"][0][0] + "";
-            venue = item["collection-title"];
-            const ISBN = item.ISBN;
-            bibtex = bibtexToString({
-                entryType: "article",
-                citationKey: doi,
-                journal: venue,
-                doi,
-                title,
-                ISBN,
-                year,
-            });
-            id = `ACM-${year}_${miniHash(doi)}`;
-            key = doi;
-            note = `Published @ ${venue} (${year})`;
-        } else {
-            throw new Error("Insufficient data from ACM citation");
-        }
-    } else {
-        throw new Error("Failed to fetch ACM citation", response);
-    }
-
-    if (venue.match(/'\d+/g)) {
-        venue = venue.replace(/'\d+/g, "");
-    }
-    if (venue.match(/\d+/g)) {
-        venue = venue.replace(/\d+/g, "");
-    }
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeIEEEPaper = async (url) => {
-    if (isPdfUrl(url)) {
-        const articleId = url
-            .split("/stamp/stamp.jsp?tp=&arnumber=")[1]
-            .match(/\d+/)[0];
-        url = `https://ieeexplore.ieee.org/document/${articleId}/`;
-    }
-    const dom = await fetchDom(url);
-    const metadata = JSON.parse(
-        [...dom.getElementsByTagName("script")]
-            .filter((s) => s.innerHTML?.includes("metadata="))[0]
-            .innerHTML.split("metadata=")[1]
-            .split(/};\s*/)[0] + "}",
-    );
-
-    const title = metadata.title;
-    const author = metadata.authors.map((a) => a.name).join(" and ");
-    const year = metadata.publicationYear;
-    const pdfLink = `${parseUrl(url).origin}${metadata.pdfUrl}`;
-    const venue = metadata.publicationTitle;
-    const key = metadata.articleId;
-    const bibtex = bibtexToString({
-        entryType: "article",
-        citationKey: key,
-        journal: venue,
-        volume: metadata.volume,
-        pages: `${metadata.startPage}-${metadata.endPage}`,
-        doi: metadata.doi,
-        title,
-        year,
-        author,
-    });
-    const id = `IEEE-${year}_${key}`;
-    const note = `Accepted @ ${venue} (${year})`;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeSpringerPaper = async (url) => {
-    // https://link.springer.com/chapter/10.1007/978-981-16-1220-6_12
-    // https://link.springer.com/article/10.1007/s00148-021-00864-z
-    // https://link.springer.com/content/pdf/10.1007/s00148-021-00864-z.pdf
-    // https://link.springer.com/article/10.1007/s00148-021-00864-z?noAccess=true
-    // https://citation-needed.springer.com/v2/references/10.1007/s41095-022-0271-y?format=bibtex&flavour=citation
-    const types = [...sourceExtras.springer.types, "content/pdf"];
-    const springerType = types.find((c) => url.includes(`/${c}/`));
-    if (!springerType) {
-        throw new Error(
-            `Could not find Springer type for ${url} (known: ${types.join(", ")})`,
-        );
-    }
-    const doi = url.split(`/${springerType}/`)[1].split("?")[0].replace(".pdf", "");
-
-    const { data } = await fetchCrossRefDataForDoi(doi);
-
-    if (!data) {
-        throw new Error("Aborting Springer paper parsing, see error above");
-    }
-
-    const { author, bibtex, citationKey, year, title, venue } = data;
-
-    const id = `Springer-${year}_${miniHash(doi)}`;
-    const note = `Published @ ${venue} (${year})`;
-
-    const pdfLink =
-        data.pdf ??
-        (springerType === "content/pdf"
-            ? url
-            : url.replace(`/${springerType}/`, "/content/pdf/") + ".pdf");
-
-    return {
-        author,
-        bibtex,
-        id,
-        key: citationKey,
-        note,
-        pdfLink,
-        title,
-        venue,
-        year,
-        extra: { url: `https://doi.org/${doi}` },
-    };
-};
-
-export const makeAPSPaper = async (url) => {
-    url = url.split("#")[0];
-    const [journal, type] = parseUrl(url).pathname.split("/").slice(1, 3);
-    const doi = url.split(`/${journal}/${type}/`).last();
-    const exportPath = url.replace(`/${journal}/${type}/`, `/${journal}/export/`);
-    const bibtex = await fetchText(`${exportPath}?type=bibtex&download=true`);
-    const data = bibtexToObject(bibtex);
-    const pdfLink = url.replace(`/${journal}/${type}/`, `/${journal}/pdf/`);
-    const id = `APS-${data.year}_${miniHash(doi)}`;
-    const journalKey = data.journal ?? data.publisher;
-    await readJournalAbbreviations();
-    const venue = journalAbbreviations[miniHash(journalKey)] ?? journalKey;
-    const note = `Published @ ${venue} (${data.year})`;
-    return {
-        author: flipAndAuthors(data.author),
-        bibtex,
-        id,
-        key: data.citationKey,
-        note,
-        pdfLink,
-        title: data.title,
-        venue,
-        year: data.year,
-    };
-};
-
-export const makeWileyPaper = async (url) => {
-    url = noParamUrl(url);
-    const pdfLink = url.match(/\/doi\/10\./g)
-        ? url.replace("/doi/", "/doi/pdf/")
-        : url.replace(/\/doi\/(abs|epdf|full)\//g, "/doi/pdf/");
-    const absLink = pdfLink.replace("/doi/pdf/", "/doi/abs/");
-    const doi = absLink.split("/doi/abs/")[1];
-    const paper = await fetchBibtexToPaper({ doi });
-    const { author, citationKey, title, year } = paper;
-    const id = `Wiley-${year}_${miniHash(doi)}`;
-    const key = citationKey;
-    const bibtex = paper.bibtex;
-    const venue = paper.journal;
-    const publisher = paper.publisher;
-    const note = `Published @ ${venue} (${year})`;
-
-    return {
-        author,
-        bibtex,
-        id,
-        key: miniHash(doi),
-        note,
-        pdfLink,
-        title,
-        venue,
-        year,
-        doi,
-    };
-};
-
-export const makeScienceDirectPaper = async (url) => {
-    const pii = url.split("/pii/")[1].split("/")[0].split("#")[0].split("?")[0];
-    const bibtex = await fetchText(
-        `https://www.sciencedirect.com/sdfe/arp/cite?pii=${pii}&format=text%2Fx-bibtex&withabstract=false`,
-    );
-    const data = bibtexToObject(bibtex);
-
-    const { author, journal, year, doi, title, citationKey } = data;
-    const note = `Published @ ${journal} (${year})`;
-    const id = `ScienceDirect-${year}_${miniHash(pii)}`;
-    const venue = journal ?? "Science Direct";
-    const pdfLink = `https://reader.elsevier.com/reader/sd/pii/${pii}`;
-
-    return {
-        author,
-        bibtex,
-        id,
-        key: citationKey,
-        note,
-        pdfLink,
-        title,
-        venue,
-        year,
-    };
-};
-
-export const makeSciencePaper = async (url) => {
-    let author, bibtex, id, key, note, pdfLink, title, venue, year, doi, absUrl;
-
-    doi = noParamUrl(url).split("/doi/")[1];
-    if (!doi.startsWith("10.")) {
-        doi = doi.split("/").slice(1).join("/");
-    }
-    pdfLink = `https://science.org/doi/pdf/${doi}`;
-    absUrl = `https://science.org/doi/full/${doi}`;
-
-    const { data } = await fetchCrossRefDataForDoi(doi);
-    if (data) {
-        ({ author, bibtex, title, venue, year } = data);
-        key = data.citationKey;
-        note = `Published @ ${venue} (${year})`;
-    } else {
-        const dom = await fetchDom(absUrl);
-        ({ author, year, publisher, title, venue, key, bibtex, note } =
-            extractDataFromDCMetaTags(dom));
-    }
-
-    id = `Science-${year}_${miniHash(doi)}`;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeFrontiersPaper = async (url) => {
-    url = url.replace(/\/pdf$/, "/full");
-    if (url.endsWith("/")) {
-        url = url.slice(0, -1);
-    }
-    const doi = noParamUrl(url).split("/articles/")[1].split("/full")[0];
-    const bib = await fetchText(noParamUrl(url).replace("/full", "") + "/bibTex");
-    const data = Object.fromEntries(
-        Object.entries(bibtexToObject(bib)).map(([k, v]) => [
-            k === "citationKey" || k === "entryType" ? k : k.toLowerCase(),
-            v,
-        ]),
-    );
-    data.author = flipAndAuthors(data.author);
-    delete data.abstract;
-    const { author, journal, year, title, citationKey } = data;
-    const bibtex = bibtexToString(data);
-
-    const venue = journal;
-    const note = `Published @ ${venue} (${year})`;
-    const id = `Frontiers-${year}_${miniHash(doi)}`;
-    const key = citationKey;
-    const pdfLink = url.replace(/\/full$/, "/pdf");
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
-export const makeIHEPPaper = async (url) => {
-    let data, num;
-    if (url.includes("/files/")) {
-        const hash = url.split("/files/")[1].split("/")[0];
-        const api = `https://inspirehep.net/api/literature?q=documents.key:${hash}`;
-        const results = (await fetchJSON(api)).data;
-        data = results.hits.hits.find(
-            (h) => !!h.metadata.documents.find((d) => d.key === hash),
-        );
-        if (!data) {
-            warn("Could not find an Inspire HEP record for the url", url);
-            return;
-        }
-        num = data.metadata.control_number;
-    } else {
-        num = url.match(/\/literature\/(\d+)/)[1];
-    }
-    if (!num) {
-        warn("Could not find an Inspire HEP id for the url", url);
-        return;
-    }
-    const bibtex = await fetchText(
-        `https://inspirehep.net/api/literature/${num}?format=bibtex`,
-    );
-    if (!data) {
-        ({ data } = await fetchJSON(
-            `https://inspirehep.net/api/literature/${num}?format=json`,
-        ));
-    }
-    const bibObj = bibtexToObject(bibtex);
-    let title = bibObj.title ?? data.metadata.titles[0].title;
-    if (title.startsWith("{") && title.endsWith("}")) title = title.slice(1, -1);
-    const pdfLink = data.metadata.documents?.[0]?.url ?? url;
-    const author = flipAndAuthors(bibObj.author);
-    const year = bibObj.year ?? data.created.split("-")[0];
-    const id = `IHEP-${num}`;
-    const venue = bibObj.journal ?? "Inspire HEP";
-    const key = bibObj.citationKey;
-    const note = `Published @ ${venue} (${year})`;
-    const doi = bibObj.doi ?? "";
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
-export const makePLOSPaper = async (url) => {
-    const doi = url.split("?id=").last().split("&")[0];
-    let { bibtex, key, author, venue, title, note, year } = await fetchBibtexToPaper({
-        doi,
-    });
-    const pdfLink = `${url.split("/article")[0]}/article/file?id=${doi}&type=printable`;
-    const section = url.split("journals.plos.org/")[1].split("/")[0];
-
-    author = flipAndAuthors(author);
-    const id = `PLOS-${section}_${miniHash(doi)}`;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
-export const makeRSCPaper = async (url) => {
-    url = noParamUrl(url).replace("/unauth", "");
-    const rscId = url.split("/").last();
-    const type = url
-        .split("/")
-        .find(
-            (s) => s === "articlehtml" || s === "articlepdf" || s === "articlelanding",
-        )
-        .replace("article", "");
-    const pdfLink =
-        type === "articlepdf" ? url : url.replace(`/article${type}/`, "/articlepdf/");
-
-    let { bibtex, key, author, venue, title, note, year, doi } =
-        await fetchBibtexToPaper({
-            url: `https://pubs.rsc.org/en/content/formatedresult?markedids=${rscId}&downloadtype=article&managertype=bibtex`,
-        });
-    author = flipAndAuthors(author);
-    const id = `RSC-${venue.replaceAll(" ", "")}_${miniHash(rscId)}`;
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
 export const parseAIPIdOrDOI = (url) => {
     if (isPdfUrl(noParamUrl(url))) {
         return {
@@ -1437,131 +437,6 @@ export const parseAIPIdOrDOI = (url) => {
               : url.split("/article-abstract/")[1].split("/")[3],
     };
 };
-export const makeAIPPaper = async (url) => {
-    url = noParamUrl(url);
-    if (isPdfUrl(url)) {
-        warn("PaperMemory cannot parse AIP papers from pdf urls");
-        return;
-    }
-    const { aipId } = parseAIPIdOrDOI(url);
-    const bibURL = `https://pubs.aip.org/Citation/Download?resourceId=${aipId}&resourceType=3&citationFormat=2`;
-    const { author, bibtex, key, note, eprint, title, venue, year, doi } =
-        await fetchBibtexToPaper({ url: bibURL });
-    const id = `AIP-${year}_${miniHash(aipId)}`;
-    const pdfLink = eprint.replaceAll("\\", "");
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
-export const makeWebsitePaper = async (tab) => {
-    const url = tab.url;
-    const dom = await fetchDom(url);
-    const og = Object.fromEntries(
-        [...dom.querySelectorAll("meta")]
-            .filter((m) => m.getAttribute("property"))
-            .filter((m) => m.getAttribute("property").startsWith("og:"))
-            .map((m) => [
-                m.getAttribute("property").replace("og:", ""),
-                m.getAttribute("content"),
-            ]),
-    );
-
-    const author =
-        og.site_name || parseUrl(url).hostname.replace("www.", "").capitalize();
-    const year = new Date().getFullYear() + "";
-    const id = `Website_${urlToWebsiteId(url)}`;
-    const note = og.description || "";
-    const pdfLink = url;
-    const title = og.title || tab.title;
-    const key = `${miniHash(author)}${year}${firstNonStopLowercase(title)}`;
-    const venue = "";
-    const accessDate = new Date().toISOString().split("T")[0];
-    const bib = `@misc{${key},
-        author = {${author}},
-        title = {${title}},
-        year = {${year}},
-        url = {${url}},
-        note = {Accessed ${accessDate}}
-    }`;
-    const bibtex = bibtexToString(bibtexToObject(bib));
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year };
-};
-
-export const makeMDPIPaper = async (url) => {
-    url = noParamUrl(url);
-    if (url.split("/").last().startsWith("pdf")) {
-        url = url.split("/").slice(0, -1).join("/");
-    }
-    if (url.endsWith("/notes")) {
-        url = url.replace("/notes", "");
-    }
-    if (url.endsWith("/reprints")) {
-        url = url.replace("/reprints", "");
-    }
-    const dom = await fetchDom(url);
-    let { author, year, publisher, title, venue, key, doi, bibtex, note, pdfLink } =
-        extractDataFromDCMetaTags(dom);
-
-    const id = `MDPI-${year}_${miniHash(url.split("mdpi.com/")[1])}`;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
-export const makeOUPPaper = async (url) => {
-    url = noParamUrl(url);
-    const resourceId = url.split("/").last();
-    let bibtex = await fetchText(
-        `https://academic.oup.com/Citation/Download?resourceId=${resourceId}&resourceType=3&citationFormat=2`,
-    );
-    const paper = bibtexToObject(bibtex);
-    delete paper.abstract;
-    bibtex = bibtexToString(paper);
-    let { title, year, author, journal, doi, citationKey, eprint } = paper;
-    author = flipAndAuthors(author);
-    const venue = journal;
-    const note = `Published @ ${venue} (${year})`;
-    const key = citationKey;
-    const num = url.split("https://academic.oup.com/")[1].split("/").slice(2).join("");
-    const id = `OUP-${year}_${miniHash(num)}`;
-    const pdfLink = eprint?.replaceAll("\\", "") ?? url;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
-export const makeHALPaper = async (url) => {
-    url = noParamUrl(url).replace(/(hal\.science\/\w+-\d+)(v\d+)?(\/document)?/, "$1"); // remove version
-    const halId = url.match(/(hal-\d+)/)[1];
-    const bibURL = `https://hal.science/${halId}/bibtex`;
-    let bibtex = await fetchText(bibURL);
-    const paper = bibtexToObject(bibtex);
-    let { title, year, journal, author, doi, pdf } = paper;
-
-    const venue = journal;
-    const note = venue ? `Published @ ${venue} (${year})` : "";
-    const key = paper.citationKey;
-    author = flipAndAuthors(author);
-    bibtex = bibtexToString(bibtex);
-    const id = `HAL-${year}_${miniHash(halId)}`;
-    const pdfLink = pdf ?? url;
-
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
-export const makeChemRxivPaper = async (url) => {
-    let chemRxivId;
-    let absUrl = url;
-    if (isPdfUrl(url)) {
-        chemRxivId = url.split("/item/")[1].split("/")[0];
-        absUrl = "https://chemrxiv.org/engage/chemrxiv/article-details/" + chemRxivId;
-    } else {
-        chemRxivId = noParamUrl(url).split("/").last();
-    }
-    const dom = await fetchDom(absUrl);
-    const { author, year, publisher, title, venue, key, doi, bibtex, pdfLink, note } =
-        extractDataFromDCMetaTags(dom);
-    const id = `ChemRxiv-${year}_${miniHash(chemRxivId)}`;
-    return { author, bibtex, id, key, note, pdfLink, title, venue, year, doi };
-};
-
 export const findCellPii = async (url) => {
     const isPdf = url.toLowerCase().includes("showpdf");
     const isPdfExtended = url.toLowerCase().includes("pdfextended");
@@ -1592,589 +467,4 @@ export const findCellPii = async (url) => {
         pii = noParamUrl(url).split("/").last();
     }
     return { pii, url };
-};
-
-export const makeCellPaper = async (url) => {
-    let pii;
-    ({ pii, url } = await findCellPii(url));
-    const pdfLink = `https://www.cell.com/action/showPdf?pii=${pii}`;
-    const dom = await fetchDom(url);
-    const doi = dom.head
-        .querySelector('meta[name="citation_doi"]')
-        .getAttribute("content");
-    const paper = await fetchBibtexToPaper({ doi });
-    const { author, year, title, venue, bibtex, note, citationKey } = paper;
-    const id = `Cell-${year}_${miniHash(url.split("cell.com/")[1])}`;
-    return {
-        author,
-        bibtex,
-        id,
-        key: citationKey,
-        note,
-        pdfLink,
-        title,
-        venue,
-        year,
-        doi,
-    };
-};
-
-// -------------------------------
-// -----  PREPRINT MATCHING  -----
-// -------------------------------
-
-export const tryPWCMatch = async (paper) => {
-    const pwcPrefs = (await getStorage("pwcPrefs")) ?? {};
-    let bibtex;
-    const payload = {
-        type: "papersWithCode",
-        pwcPrefs,
-        paper: paper,
-    };
-    const { url, note, venue, pubYear } =
-        (await sendMessageToBackground(payload)) ?? {};
-    if (url && !paper.codeLink) {
-        log("[PapersWithCode] Discovered a code repository:", url);
-    } else {
-        log("[PapersWithCode] No code repository found");
-    }
-    if (venue && !paper.venue) {
-        log("[PapersWithCode] Found a publication venue:", venue);
-        const paperBib = bibtexToObject(paper.bibtex);
-        bibtex = bibtexToString({
-            ...paperBib,
-            year: pubYear,
-            journal: venue,
-        });
-    } else if (!paper.venue) {
-        log("[PapersWithCode] No publication found");
-    }
-
-    return {
-        codeLink: url,
-        note,
-        venue,
-        bibtex,
-    };
-};
-
-// --------------------------------------------
-// -----  Try CrossRef's API for a match  -----
-// --------------------------------------------
-
-/**
- * Looks for a title in crossref's database, querying titles and looking for an exact match. If no
- * exact match is found, it will return an empty note "". If a match is found and `item.event.name`
- * exists, it will be used for a new note.
- * @param {object} paper The paper to look for in crossref's database for an exact title match
- * @returns {string} The note for the paper as `Accepted @ ${items.event.name} -- [crossref.org]`
- */
-export const tryCrossRef = async (paper, toBackground) => {
-    if (toBackground) {
-        return await sendMessageToBackground({ type: "try-cross-ref", paper });
-    }
-    try {
-        // fetch crossref' api for the paper's title
-        const title = encodeURI(paper.title);
-        const api = `https://api.crossref.org/works?rows=1&mailto=schmidtv%40mila.quebec&select=event%2Ctitle&query.title=${title}`;
-        const json = await fetch(api).then((response) => response.json());
-
-        // assert the response is valid
-        if (json.status !== "ok") {
-            log(`[Crossref] ${api} returned ${json.message.status}`);
-            return {};
-        }
-        // assert there is a (loose) match
-        if (json.message.items.length === 0) return {};
-
-        // compare matched item's title to the paper's title
-        const crossTitle = json.message.items[0].title[0]
-            ?.toLowerCase()
-            .replaceAll("\n", " ")
-            .replaceAll(/\s\s+/g, " ");
-        const refTitle = paper.title
-            .toLowerCase()
-            .replaceAll("\n", " ")
-            .replaceAll(/\s\s+/g, " ");
-        if (crossTitle !== refTitle) {
-            return {};
-        }
-
-        // assert the matched item has an event with a name
-        // (this may be too restrictive for journals, to improve)
-        if (!json.message.items[0].event || !json.message.items[0].event.name) {
-            return {};
-        }
-
-        // return the note
-        info("Found a CrossRef match");
-        const venue = json.message.items[0].event.name.trim();
-        const note = `Accepted @ ${venue} -- [crossref.org]`;
-        return { venue, note };
-    } catch (error) {
-        // something went wrong, log the error, return {}
-        logError("[Crossref]", error);
-        return {};
-    }
-};
-
-export const tryDBLP = async (paper, toBackground) => {
-    if (toBackground) {
-        return await sendMessageToBackground({ type: "try-dblp", paper });
-    }
-    try {
-        const title = encodeURI(paper.title);
-        const api = `https://dblp.org/search/publ/api?q=${title}&format=json`;
-        const response = await fetch(api);
-
-        if (response.status === 429) {
-            return { status: 429 };
-        }
-
-        const json = await response.json();
-
-        if (
-            !json.result ||
-            !json.result.hits ||
-            !json.result.hits.hit ||
-            !json.result.hits.hit.length
-        ) {
-            return {};
-        }
-
-        const hits = json.result.hits.hit.sort(
-            (a, b) => parseInt(a.info.year, 10) - parseInt(b.info.year, 10),
-        );
-
-        for (const hit of hits) {
-            const hitTitle = decodeHtml(
-                hit.info.title
-                    ?.toLowerCase()
-                    .replaceAll("\n", " ")
-                    .replaceAll(".", "")
-                    .replaceAll(/\s\s+/g, " "),
-            );
-            const refTitle = paper.title
-                .toLowerCase()
-                .replaceAll("\n", " ")
-                .replaceAll(".", "")
-                .replaceAll(/\s\s+/g, " ");
-            if (hitTitle === refTitle && hit.info.venue !== "CoRR") {
-                info("Found a DBLP match");
-                const bibtex = await fetchText(hit.info.url + ".bib");
-                const abbr = miniHash(hit.info.venue);
-                await readJournalAbbreviations();
-                const venue = (journalAbbreviations[abbr] ?? hit.info.venue).trim();
-                const year = hit.info.year;
-                const url = hit.info.url;
-                const note = `Accepted @ ${venue} ${year} -- [dblp.org]`;
-                return { venue, note, bibtex };
-            }
-        }
-        return {};
-    } catch (error) {
-        // something went wrong, log the error, return {}
-        logError("[DBLP]", error);
-        return {};
-    }
-};
-
-export const trySemanticScholar = async (paper, toBackground) => {
-    if (toBackground) {
-        return await sendMessageToBackground({
-            type: "try-semantic-scholar",
-            paper,
-        });
-    }
-    try {
-        const { data, status } = await fetchJSON(
-            `https://api.semanticscholar.org/graph/v1/paper/search?query=${encodeURI(
-                paper.title,
-            )}&fields=title,venue,year,authors,externalIds,url&limit=5`,
-        );
-        const matches = data;
-
-        if (matches && matches.data && matches.data.length > 0) {
-            for (const match of matches.data) {
-                if (
-                    miniHash(match.title) === miniHash(paper.title) &&
-                    Math.abs(match.year - paper.year) < 3 &&
-                    match.venue &&
-                    !match.venue.toLowerCase().includes("arxiv") &&
-                    !match.venue.toLowerCase().includes("biorxiv")
-                ) {
-                    info("Found a Semantic Scholar match");
-                    let venue = match.venue
-                        .trim()
-                        .replace(/^\d{4}/, "")
-                        .trim();
-                    if (venue.indexOf(" ") < 0) venue = venue.toUpperCase();
-                    const year = (match.year + "").trim();
-                    const note = `Accepted @ ${venue} (${year}) -- [semanticscholar.org]`;
-                    const authors = match.authors.map((a) => a.name).join(" and ");
-                    let doi = match.externalIds.DOI;
-                    // if (doi) {
-                    //     doi = doi.replaceAll("_", "\\{_}");
-                    // }
-                    const bibtex = bibtexToString({
-                        entryType: "article",
-                        citationKey:
-                            miniHash(match.authors[0].name.split(" ").last()) +
-                            year +
-                            firstNonStopLowercase(paper.title),
-                        title: paper.title,
-                        author: authors,
-                        journal: venue,
-                        year,
-                        doi,
-                        bibSource: `Semantic Scholar ${match.url}`,
-                    });
-                    return { venue, note, bibtex, status };
-                }
-            }
-        }
-        return { status };
-    } catch (error) {
-        logError("[SemanticScholar]", error);
-    }
-    return { status: 404 };
-};
-
-export const tryGoogleScholar = async (paper) => {
-    const resp = await sendMessageToBackground({
-        type: "google-scholar",
-        paper,
-    });
-    resp.note && info("Found a Google Scholar match", resp.note);
-    return resp;
-};
-
-export const tryUnpaywall = async (paper, toBackground) => {
-    if (toBackground) {
-        return await sendMessageToBackground({ type: "try-unpaywall", paper });
-    }
-    const url = `https://api.unpaywall.org/v2/search?query=${encodeURI(
-        paper.title,
-    )}&is_oa=true&email=papermemory+${parseInt(Math.random() * 1000)}@gmail.com`;
-    const { data, status } = await fetchJSON(url);
-    if (data && status === 200) {
-        const match = data.results?.find(
-            (m) => miniHash(m.response.title) === miniHash(paper.title),
-        );
-        if (match && match.journal_name) {
-            const venue = match.journal_name;
-            const note = `Accepted @ ${venue} (${match.year}) -- [unpaywall.org]`;
-            const doi = match.doi;
-            return { venue, note, doi };
-        }
-    }
-    return { status };
-};
-
-export const tryPreprintMatch = async (paper, tryPwc = false) => {
-    let note, venue, bibtex, code, doi;
-    let matches = {};
-
-    let names = ["GoogleScholar", "SemanticScholar", "CrossRef", "DBLP", "Unpaywall"];
-    let matchPromises = [
-        silentPromiseTimeout(tryGoogleScholar(paper)),
-        silentPromiseTimeout(trySemanticScholar(paper, true)),
-        silentPromiseTimeout(tryCrossRef(paper)),
-        silentPromiseTimeout(tryDBLP(paper)),
-        silentPromiseTimeout(tryUnpaywall(paper)),
-    ];
-
-    if (tryPwc) {
-        matchPromises.push(silentPromiseTimeout(tryPWCMatch(paper)));
-        names.push("PapersWithCode");
-    }
-
-    for (const [n, name] of Object.entries(names)) {
-        matches[name] = await matchPromises[n];
-        ({ note, venue, bibtex, doi } = matches[name] ?? {});
-        if (note) {
-            break;
-        } else {
-            log(`[${name}] No publication found`);
-        }
-    }
-
-    if (tryPwc) {
-        const name = "PapersWithCode";
-        if (!matches.hasOwnProperty(name)) {
-            matches[name] = await silentPromiseTimeout(matchPromises[name]);
-        }
-        if (matches[name].codeLink && !paper.codeLink) {
-            code = matches[name].codeLink;
-        }
-    }
-
-    return { note, venue, bibtex, code, doi };
-};
-
-// -----------------------------
-// -----  Creating papers  -----
-// -----------------------------
-
-export const initPaper = async (paper) => {
-    if (!paper.note) {
-        paper.note = "";
-    }
-
-    paper.md = `[${paper.title}](${paper.pdfLink})`;
-    paper.tags = [];
-    paper.codeLink = paper.codeLink ?? "";
-    paper.favorite = false;
-    paper.favoriteDate = "";
-    paper.addDate = new Date().toJSON();
-    paper.lastOpenDate = paper.addDate;
-    paper.count = 1;
-    paper.code = {};
-    // Sanitize parsed bibtex (e.g. strip abstracts) before it is ever
-    // persisted alongside the paper.
-    if (paper.bibtex) {
-        const bibObj = sanitizeBibtexObject(bibtexToObject(paper.bibtex));
-        paper.bibtex = bibtexToString(bibObj);
-        paper.doi = paper.doi ?? bibObj.doi ?? "";
-    } else {
-        paper.doi = paper.doi ?? "";
-    }
-    for (const k in paper) {
-        if (paper.hasOwnProperty(k) && typeof paper[k] === "string") {
-            paper[k] = paper[k].trim();
-        }
-    }
-
-    paper = await autoTagPaper(paper);
-    validatePaper(paper);
-
-    return paper;
-};
-
-export const autoTagPaper = async (paper) => {
-    try {
-        const autoTags = await getStorage("autoTags");
-        if (!autoTags || !autoTags.length) return paper;
-        let tags = new Set();
-        for (const at of autoTags) {
-            if (!at.tags?.length) continue;
-            if (!at.title && !at.author) continue;
-
-            let titleMatch = true;
-            let authorMatch = true;
-            try {
-                if (at.title) {
-                    titleMatch = new RegExp(at.title, "i").test(paper.title);
-                }
-                if (at.author) {
-                    authorMatch = new RegExp(at.author, "i").test(paper.author);
-                }
-            } catch (e) {
-                continue;
-            }
-
-            if (titleMatch && authorMatch) {
-                at.tags.forEach((t) => tags.add(t));
-            }
-        }
-        paper.tags = [...tags].sort();
-        if (paper.tags.length) {
-            log("Automatically adding tags:", paper.tags);
-        }
-        return paper;
-    } catch (error) {
-        log("Error auto-tagging:", error);
-        log("Paper:", paper);
-        return paper;
-    }
-};
-
-export const makePaper = async (is, url, tab = false) => {
-    let paper;
-    let start = performance.now();
-    info("Making paper...");
-    try {
-        if (tab) {
-            paper = await makeWebsitePaper(tab);
-            if (paper) {
-                paper.source = "website";
-            }
-        } else if (is.arxiv) {
-            paper = await makeArxivPaper(url);
-            if (paper) {
-                paper.source = "arxiv";
-            }
-            // paper.codes = await fetchCodes(paper)
-        } else if (is.neurips) {
-            paper = await makeNeuripsPaper(url);
-            if (paper) {
-                paper.source = "neurips";
-            }
-            // paper.codes = await fetchCodes(paper);
-        } else if (is.cvf) {
-            paper = await makeCVFPaper(url);
-            if (paper) {
-                paper.source = "cvf";
-            }
-        } else if (is.openreview) {
-            paper = await makeOpenReviewPaper(url);
-            if (paper) {
-                paper.source = "openreview";
-            }
-        } else if (is.biorxiv) {
-            paper = await makeBioRxivPaper(url);
-            if (paper) {
-                paper.source = "biorxiv";
-            }
-        } else if (is.pmlr) {
-            paper = await makePMLRPaper(url);
-            if (paper) {
-                paper.source = "pmlr";
-            }
-        } else if (is.acl) {
-            paper = await makeACLPaper(url);
-            if (paper) {
-                paper.source = "acl";
-            }
-        } else if (is.pnas) {
-            paper = await makePNASPaper(url);
-            if (paper) {
-                paper.source = "pnas";
-            }
-        } else if (is.nature) {
-            paper = await makeNaturePaper(url);
-            if (paper) {
-                paper.source = "nature";
-            }
-        } else if (is.acs) {
-            paper = await makeACSPaper(url);
-            if (paper) {
-                paper.source = "acs";
-            }
-        } else if (is.iop) {
-            paper = await makeIOPPaper(url);
-            if (paper) {
-                paper.source = "iop";
-            }
-        } else if (is.jmlr) {
-            paper = await makeJMLRPaper(url);
-            if (paper) {
-                paper.source = "jmlr";
-            }
-        } else if (is.pmc) {
-            paper = await makePMCPaper(url);
-            if (paper) {
-                paper.source = "pmc";
-            }
-        } else if (is.ijcai) {
-            paper = await makeIJCAIPaper(url);
-            if (paper) {
-                paper.source = "ijcai";
-            }
-        } else if (is.acm) {
-            paper = await makeACMPaper(url);
-            if (paper) {
-                paper.source = "acm";
-            }
-        } else if (is.ieee) {
-            paper = await makeIEEEPaper(url);
-            if (paper) {
-                paper.source = "ieee";
-            }
-        } else if (is.springer) {
-            paper = await makeSpringerPaper(url);
-            if (paper) {
-                paper.source = "springer";
-            }
-        } else if (is.aps) {
-            paper = await makeAPSPaper(url);
-            if (paper) {
-                paper.source = "aps";
-            }
-        } else if (is.wiley) {
-            paper = await makeWileyPaper(url);
-            if (paper) {
-                paper.source = "wiley";
-            }
-        } else if (is.sciencedirect) {
-            paper = await makeScienceDirectPaper(url);
-            if (paper) {
-                paper.source = "sciencedirect";
-            }
-        } else if (is.science) {
-            paper = await makeSciencePaper(url);
-            if (paper) {
-                paper.source = "science";
-            }
-        } else if (is.frontiers) {
-            paper = await makeFrontiersPaper(url);
-            if (paper) {
-                paper.source = "frontiers";
-            }
-        } else if (is.ihep) {
-            paper = await makeIHEPPaper(url);
-            if (paper) {
-                paper.source = "ihep";
-            }
-        } else if (is.plos) {
-            paper = await makePLOSPaper(url);
-            if (paper) {
-                paper.source = "plos";
-            }
-        } else if (is.rsc) {
-            paper = await makeRSCPaper(url);
-            if (paper) {
-                paper.source = "rsc";
-            }
-        } else if (is.mdpi) {
-            paper = await makeMDPIPaper(url);
-            if (paper) {
-                paper.source = "mdpi";
-            }
-        } else if (is.oup) {
-            paper = await makeOUPPaper(url);
-            if (paper) {
-                paper.source = "oup";
-            }
-        } else if (is.hal) {
-            paper = await makeHALPaper(url);
-            if (paper) {
-                paper.source = "hal";
-            }
-        } else if (is.chemrxiv) {
-            paper = await makeChemRxivPaper(url);
-            if (paper) {
-                paper.source = "chemrxiv";
-            }
-        } else if (is.cell) {
-            paper = await makeCellPaper(url);
-            if (paper) {
-                paper.source = "cell";
-            }
-        } else if (is.aip) {
-            paper = await makeAIPPaper(url);
-            if (paper) {
-                paper.source = "aip";
-            }
-        } else {
-            console.error({ is, url });
-            throw new Error(
-                "Could not parse paper (in `makePaper`). Unknown paper source, see above.",
-            );
-        }
-    } catch (e) {
-        logError("Error in makePaper:", e);
-        if (e.message?.includes("Unknown paper source")) {
-            throw e;
-        }
-        return;
-    }
-
-    if (typeof paper === "undefined") {
-        return;
-    }
-    const elapsed = (performance.now() - start) / 1000;
-    info(`Paper parsed in ${elapsed.toFixed(2)}s`);
-
-    return await initPaper(paper);
 };
